# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/password/forgot": {
      "post": {
        "description": "Always answers 202, whether the email exists or not (it never reveals accounts). If it exists, emails a single-use link to PASSWORD_RESET_URL?token=… that expires in minutes. Limited per email and per IP (429).",
        "operationId": "AuthController_forgot_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "If the account exists, the email is on its way"
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Forgot my password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/password/reset": {
      "post": {
        "description": "With the token from the email. The link works once; afterwards every session is closed (log in with the new password).",
        "operationId": "AuthController_reset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Reset my password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "description": "Role `user`. The birth date decides access to alcoholic drinks (18+).",
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with existing data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Create an account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "description": "Returns a short-lived access token. By default the refresh token is set as an httpOnly cookie (send requests with credentials); ask for refreshTokenIn: \"body\" if you are not a browser. Repeated failures lock the account and the IP for a while (429 LOGIN_LOCKED with Retry-After).",
        "operationId": "AuthController_signIn_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. See the Retry-After header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Log in",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Uses the refresh token from the body or, if omitted, from the session cookie (only from the allowed frontend origins). The token rotates on every use; reusing an old one revokes every session of that login.",
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Refresh the session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Revokes the session (token from the body or the cookie) and clears the cookie. Idempotent.",
        "operationId": "AuthController_signOut_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Logged out"
          },
          "400": {
            "description": "Invalid input (params, query or body).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Log out",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My profile",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "api-drinks",
    "description": "\nAPI de bebidas y cocteles sobre TheCocktailDB: despensa inteligente, cocteles por estado de ánimo,\nADN de sabor, gemelos y carta inteligente para bares.\n\n**Versión**: todas las rutas van bajo `/v1`, salvo `/health`, `/health/ready` y `/`.\n\n**Errores**: todos tienen la forma `{ statusCode, code, message, error, details? }`. El frontend decide por `code`\n(estable) y muestra su propio texto; `message` es solo una pista en inglés.\n\n**Autenticación**\n- Apps propias: `Authorization: Bearer <accessToken>` (ver /v1/auth/login). En navegador, el refresh token\n  viaja en una cookie httpOnly: haz las peticiones de /v1/auth con credentials: 'include'.\n- Integraciones de terceros: `X-API-Key: <key>` (ver /v1/me/api-keys). Cuentan contra la cuota diaria del plan\n  y responden con los headers `X-RateLimit-Limit`, `X-RateLimit-Remaining` y `X-RateLimit-Reset`.\n\n**Fuente de datos**: recetas e imágenes de [TheCocktailDB](https://www.thecocktaildb.com/). Todo frontend que\nmuestre bebidas debe mostrar \"Datos e imágenes: TheCocktailDB\" con un enlace a su sitio (condición de su uso).\n\n**Verificación de edad**: las bebidas con alcohol solo se muestran a usuarios autenticados mayores de 18 años.\n",
    "version": "4.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "email": {
            "type": "string",
            "example": "ana@example.com"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "premium",
              "bartender",
              "venue_owner",
              "admin"
            ]
          },
          "birthDate": {
            "type": "string",
            "example": "1995-05-20"
          },
          "adult": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "birthDate",
          "adult",
          "createdAt"
        ],
        "additionalProperties": false,
        "description": "Public view of a user. Never includes the password hash."
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "example": 403
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "PLAN_LIMIT",
              "RATE_LIMITED",
              "UNAVAILABLE",
              "VALIDATION_FAILED",
              "ROUTE_NOT_FOUND",
              "PAYLOAD_TOO_LARGE",
              "INTERNAL_ERROR",
              "AUTH_REQUIRED",
              "INVALID_CREDENTIALS",
              "INVALID_ACCESS_TOKEN",
              "INVALID_REFRESH_TOKEN",
              "INVALID_API_KEY",
              "API_KEY_NOT_ALLOWED",
              "AMBIGUOUS_CREDENTIALS",
              "ORIGIN_NOT_ALLOWED",
              "ROLE_REQUIRED",
              "AGE_RESTRICTED",
              "LOGIN_LOCKED",
              "API_QUOTA_EXCEEDED",
              "EMAIL_TAKEN",
              "INVALID_EMAIL",
              "WEAK_PASSWORD",
              "INVALID_BIRTH_DATE",
              "ADMIN_SELF_DEMOTION",
              "ADMIN_SELF_DELETION",
              "WRONG_PASSWORD",
              "INVALID_RESET_TOKEN",
              "INVALID_NAME",
              "USER_NOT_FOUND",
              "DRINK_NOT_FOUND",
              "MOOD_NOT_FOUND",
              "CATALOG_UNAVAILABLE",
              "NO_TASTE_YET",
              "TASTE_LINK_NOT_FOUND",
              "OWN_TASTE_LINK",
              "GAME_OVER",
              "DUPLICATE_GUESS",
              "PLAN_NOT_FOUND",
              "PLAN_NOT_PURCHASABLE",
              "PAYMENTS_UNAVAILABLE",
              "PAYMENT_NOT_FOUND",
              "INVALID_WEBHOOK_SIGNATURE",
              "VENUE_NOT_FOUND",
              "NOT_VENUE_OWNER"
            ],
            "description": "Stable error code: switch on it and show your own (translated) text",
            "example": "AGE_RESTRICTED"
          },
          "message": {
            "type": "string",
            "description": "Human hint in English; may change, do not parse it",
            "example": "This drink contains alcohol: sign in with a verified adult account to see it"
          },
          "error": {
            "type": "string",
            "example": "ForbiddenError"
          },
          "details": {
            "description": "Only with VALIDATION_FAILED: one entry per invalid field",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": false
            }
          },
          "requestId": {
            "description": "Same as the X-Request-Id header: quote it when reporting a problem",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "error"
        ],
        "additionalProperties": false,
        "description": "Every error response has this shape."
      },
...
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "example": "ana@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "description": "The token from the emailed link (?token=…)"
          },
          "newPassword": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "10+ characters with letters and numbers"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "example": "ana@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "10+ characters with letters and numbers",
            "example": "secret-password-123"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Ana"
          },
          "birthDate": {
            "description": "YYYY-MM-DD. Adults (18+) can see alcoholic drinks.",
            "example": "1995-05-20",
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "birthDate"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "example": "ana@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "example": "secret-password-123"
          },
          "refreshTokenIn": {
            "default": "cookie",
            "type": "string",
            "enum": [
              "cookie",
              "body"
            ],
            "description": "Where to deliver the refresh token. cookie (default): httpOnly cookie scoped to /v1/auth, unreadable by JavaScript. body: in the JSON, for clients without a browser."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "Session": {
        "type": "object",
        "properties": {
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "accessToken": {
            "type": "string",
            "description": "JWT. Send as Authorization: Bearer <token>"
          },
          "expiresIn": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Access token lifetime in seconds"
          },
          "refreshToken": {
            "nullable": true,
            "description": "Opaque, single use (each refresh returns a new one). null when it was set as an httpOnly cookie.",
            "type": "string"
          },
          "refreshTokenIn": {
            "type": "string",
            "enum": [
              "cookie",
              "body"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "tokenType",
          "accessToken",
          "expiresIn",
          "refreshToken",
          "refreshTokenIn",
          "user"
        ],
        "additionalProperties": false
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "description": "Omit it to use the session cookie",
            "type": "string",
            "minLength": 1
          },
          "refreshTokenIn": {
            "description": "Default: the same place the token came from",
            "type": "string",
            "enum": [
              "cookie",
              "body"
            ]
          }
        }
      },
...
    }
  }
}